requirements.txt
Django==3.2
djangorestframework==3.13.1
requests==2.27.1